requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.7.2
langchain-core==0.1.52
langchain-community==0.0.38
langchain-mongodb==0.1.0
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.2
google-generativeai==0.5.4
sentence-transformers==2.7.0
pypdf==4.2.0
python-docx==1.1.2